EqualizerTile: Remove shade lag during swipe to flip when enabled
Linking and unlinking the visualizer to the view every time the shade state changed back and forth
was causing some lag at the halfway point-- so introduce a slight delay as a bandaid to allow
animations and methods being called during the flip to finish processing before the EQ tile updates it's
link state.  This delay was observed on my 1+1 running the stock CMS11.0 also.

Change-Id: Ib293995cad5e793c232d851be81d5810f690bd78